Getting information from JSON
Created if statements to check a player's coordinates and get an id based off of the coordinate. Created a getEncounters function that makes an ajax call and returns the index of the json object that correlates to the id of the encounter. Set up an api route to send information to the front end.